changed price from int to double

+++ menu_item.h
@@ -6,18 +6,18 @@ using namespace std;
 class menu_item {
 protected:
     string item_name;
-    int item_price;
+    double item_price;
     int item_size;
 public:
     menu_item();
-    menu_item(string name, int price, int size);
+    menu_item(string name, double price, int size);
     
     void set_item_name(string name);
-    void set_item_price(int price);
+    void set_item_price(double price);
     void set_item_size(int size);
 
     string get_item_name();
-    int get_item_price();
+    double get_item_price();
     int get_item_size();
 };
 
@@ -29,7 +29,7 @@ menu_item::menu_item() {
 }
 
 // Menu_item constructor
-menu_item::menu_item(string name, int price, int size) {
+menu_item::menu_item(string name, double price, int size) {
     item_name = name;
     item_price = price;
     item_size = size;
@@ -40,7 +40,7 @@ void menu_item::set_item_name(string name) {
     item_name = name;
     cout << "This item's name is changed to " << item_name << endl;
 }
-void menu_item::set_item_price(int price) {
+void menu_item::set_item_price(double price) {
     item_price = price;
     cout << "This item's price is changed to " << item_price << endl;
 }
@@ -53,7 +53,7 @@ void menu_item::set_item_size(int size) {
 string menu_item::get_item_name() {
     return item_name;
 }
-int menu_item::get_item_price() {
+double menu_item::get_item_price() {
     return item_price;
 }
 int menu_item::get_item_size() {
